🚚 move: relocate product-media-list related files to /bridge/productmedialist package
