+++ src/main/java/com/jokim/sivillage/api/bridge/productmedialist/application/ProductMediaListService.java
@@ -1,7 +1,7 @@
-package com.jokim.sivillage.api.bridge.application;
+package com.jokim.sivillage.api.bridge.productmedialist.application;
 
-import com.jokim.sivillage.api.bridge.dto.in.ProductMediaListRequestDto;
-import com.jokim.sivillage.api.bridge.dto.out.ProductMediaListResponseDto;
+import com.jokim.sivillage.api.bridge.productmedialist.dto.in.ProductMediaListRequestDto;
+import com.jokim.sivillage.api.bridge.productmedialist.dto.out.ProductMediaListResponseDto;
 
 import java.util.List;
